data_pipeline/requirements.txt
pyarrow==19.0.1
geopandas==1.0.1
requests==2.32.3
rasterio==1.4.3
boto3==1.36.22
numpy==2.2.3
pandas==2.2.3
tqdm==4.67.1
python-dotenv==1.0.1
PyYAML==6.0.2
